Overrode default project.nvim autoroot mechanic
Autoroot is disabled (manual = true), and a custom function that calls the autoroot
command directly has been made in its place, with the exception that it checks if
a global variable is set or not to determine whether or not to execute the command.

A new command has also been created to toggle this global on and off, which allows
for the easy enabling/disabling of the autoroot system, whenever needed.