A simple module to check how Neovim was launched
Checks if Neovim was run without any arguments, or if a file was provided.
The Alpha plugin (the greeter I use) depends on this, to know whether or
not it should lazy load. It doesn't make much sense to load a greeter
plugin if you're never going to see it, since you just opened a file
directly.