Custom header/source switch because clangd-extensions' is scuffed.
The problem with clangd-extensions' implementation is that it only looks for the
matching header in the same directory as the source file, and has no fallback
to look through open buffers instead if a match isn't found.

This is annoying when you structure your C++ project in a way that separates
headers and source files into separate directories.

Currently not operatoinal, just scaffolding for the most part. Work in progress.
This implementation should (configurably) priioritize the current folder by
default, and fall back to looking through opened buffers, by default anyway.