Added flashbang.lua
I created this because I was sick of getting flashbanged while browsing through colorschemes. This tests every schemes and checks if it's dark or light respectively, then caches the results.

Now, Coldark and Colight user commands exist to select only from the list of dark or light themes.
Flashbang is the main command to reload or sync the cache with the schemes currently loaded.

(note: every colorscheme has to be loaded, so setting everything to not lazy load temporarily, to make it visible to sync, then re-enabling it is the way to go)
(note, note: if you sync again with those schemes not loaded, they will be removed from the cache, so don't sync frequently)

Autocmd should add not seen before clorschemes to the cache automatically anyway.

TODO:
- Forgot to classify "other" (no guibg set).
- Write a telescope picker.