Disabled style=minimal and made it minimal manually
I repeated the same beahvior documented by setting the style to minimal,
except the very last part, because changing the window highlights before
a colorscheme gets loaded anyway causes it to flash momentarily, which
looks ugly. That does mean that the highlights are weird when using
default theme, will have to look into fixing that/timing the load order.